+++ backend/Services/Interfaces/Card/ICardFacePerCardService.cs
@@ -5,9 +5,9 @@ namespace Services
 {
     public interface ICardFacePerCardService: ICrud<CardFacePerCard>, ICrudNav<CardFacePerCard>
     {
-        public Task<IEnumerable<CardFacePerCard>> GetAllNavByCardId(int cardId);
+        public Task<IEnumerable<CardFacePerCard>> GetAllNavByCardId(long cardId);
 
-        public Task<IEnumerable<CardFacePerCard>> GetAllByCardId(int cardId);
+        public Task<IEnumerable<CardFacePerCard>> GetAllByCardId(long cardId);
     
         public Task CreateAsyncFromCardEditorCardDto(CardEditorCardDto cardEditorCardDto);
     }
