+++ backend/Services/Interfaces/Card/ICardPerOwnerService.cs
@@ -6,11 +6,11 @@ namespace Services
 {
     public interface ICardPerOwnerService: ICrud<CardPerOwner>
     {
-        public Task<CardPerOwner?> GetNavByCardIdAsync(int cardId);
+        public Task<CardPerOwner?> GetNavByCardIdAsync(long cardId);
 
-        public Task<CardPerOwner?> GetByCardIdAndOwnerIdAsync(int cardId, string ownerId);
+        public Task<CardPerOwner?> GetByCardIdAndOwnerIdAsync(long cardId, string ownerId);
 
-        public Task<CardPerOwner?> GetByCardIdAsync(int cardId);
+        public Task<CardPerOwner?> GetByCardIdAsync(long cardId);
 
         public Task<IEnumerable<Card>> GetCardsNavByOwnerIdAsync(string ownerId);
     }
