+++ backend/Services/Interfaces/General/CRUD/ICrudDto.cs
@@ -2,9 +2,9 @@ namespace Services
 {
     public interface ICrudDto<T>
     {
-        public Task<T?> GetDtoAsync(int id);
+        public Task<T?> GetDtoAsync(long id);
         public Task CreateDtoAsync(T dto);
-        public Task<bool> UpdateDtoAsync(int id, T dto);
-        public Task<bool> DeleteDtoAsync(int id);
+        public Task<bool> UpdateDtoAsync(long id, T dto);
+        public Task<bool> DeleteDtoAsync(long id);
     }
 }
